moved location availability check to LocationService, references #45
